+++ src/controller/AppController.ts
@@ -2,13 +2,14 @@ import WeightManager from '@/core/WeightManager.ts';
 import NeuralNetworkBase from '@/core/NeuralNetworkBase.ts';
 import QueryProcessController from './queryPipeline/QueryProcessController.ts';
 import PerceptronController from '@/controller/perceptron/PerceptronController.ts';
+import BaseCanvas from '@/view/components/perceptron/BaseCanvas.ts';
 import UserInputCanvas from '@/view/components/userQuery/UserInputCanvas.ts';
 import PathTrackingCanvas from '@/view/components/userQuery/PathTrackingCanvas.ts';
 import AlignCanvas from '@/view/components/userQuery/AlignCanvas.ts';
 import ResizeCanvas from '@/view/components/userQuery/ResizeCanvas.ts';
 import DataStore from './DataStore.ts';
 
-import PerceptronController from '@/controller/perceptron/PerceptronController.ts';
+import NodeRenderer from '@/view/components/perceptron/NodeRenderer.ts';
 
 import eventBus from './EventBus.ts';
 import { DATA_EVENTS, HANDLER_EVENTS } from './constants/events.ts';
@@ -18,8 +19,8 @@ import { IWeightManager } from '@/core/types/WeightManager.ts';
 import { INeuralNetworkBase } from '@/core/types/NeuralNetworkBase.ts';
 import { IDataStore } from './types/DataStore.ts';
 import { Matrix2D } from '@/core/ops/types/OpsType.ts';
-import { NodePositionHandler } from '@/controller/perceptron/NodePositionHandler.ts';
-import { EdgePositionHandler } from '@/controller/perceptron/EdgePositionHandler.ts';
+import { NodeHandler } from '@/controller/perceptron/NodeHandler.ts';
+import { EdgePositionHandler } from '@/controller/perceptron/EdgeHandler.ts';
 import { ScrollEventHandler } from '@/controller/perceptron/ScrollEventHandler.ts';
 
 class AppController {
@@ -40,8 +41,16 @@ class AppController {
             $NN: $NN,
             $DS: $DS,
         });
+
+        const perceptronBaseCanvas = new BaseCanvas();
+        const nodeRenderer = new NodeRenderer(perceptronBaseCanvas.getCtx());
+        const nodeHandler = new NodeHandler({
+            networkConfig: NETWORK_CONFIG,
+            nodeRenderer,
+        });
         const $PC = new PerceptronController({
-            nodePositionHandler: new NodePositionHandler(NETWORK_CONFIG),
+            nodeRenderer: nodeRenderer,
+            nodePositionHandler: nodeHandler,
             edgePositionHandler: new EdgePositionHandler(),
             scrollEventHandler: new ScrollEventHandler(),
         });
