+++ src/controller/constants/networkConfig.ts
@@ -4,5 +4,4 @@ export const NETWORK_CONFIG: networkConfig = {
     inputNodes: 784,
     hiddenNodes: 200,
     outputNodes: 10,
-    learningRate: 0.15,
 };
