+++ src/controller/constants/types/networkConfig.ts
@@ -2,5 +2,4 @@ export interface networkConfig {
     inputNodes: number;
     hiddenNodes: number;
     outputNodes: number;
-    learningRate: number;
 }
