--- src/controller/perceptron/NodePositionHandler.ts
@@ -1,9 +1,0 @@
-import { networkConfig } from '@/controller/constants/types/networkConfig.ts';
-
-export class NodePositionHandler {
-    private networkConfig: networkConfig;
-
-    constructor(networkConfig: networkConfig) {
-        this.networkConfig = networkConfig;
-    }
-}
