+++ src/controller/perceptron/types/PerceptronController.ts
@@ -1,5 +1,6 @@
 export interface IPerceptronControllerProps {
     nodePositionHandler: INodePositionHandler;
     edgePositionHandler: INdgePositionHandler;
+    nodeRenderer: INodeRenderer;
     scrollEventHandler: IScrollEventHandler;
 }
