+++ src/view/components/perceptron/types/Node.ts
@@ -1,12 +1,1 @@
-export type layer = string;
-export type index = number;
-
-export interface Position {
-    x: number;
-    y: number;
-}
-
-export interface edgePosition {
-    x: number;
-    y: number;
-}
+export type value = number;
